requirements.txt
python-dotenv==1.0.0
gTTS
telebot